Reorganize kodi entry point
